LPD-56913 Rename fragment classes to match superclass
