+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layouts/01_dashboard/page-definition.json
@@ -5,7 +5,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ViewDashboardFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ViewDashboardJSPSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
