+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layouts/02_assets/02_contents/page-definition.json
@@ -5,7 +5,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ViewContentsSectionFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ViewContentsJSPSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
