+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layouts/02_assets/03_files/page-definition.json
@@ -5,7 +5,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ViewFilesSectionFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ViewFilesJSPSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
