+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layouts/03_admin/01_structures/01_structure_builder/page-definition.json
@@ -5,7 +5,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.StructureBuilderFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.StructureBuilderComponentSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
