+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layouts/03_admin/03_categorization/01_vocabularies/01_add_vocabulary/page-definition.json
@@ -4,7 +4,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.EditVocabularyFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.EditVocabularyJSPSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
