+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layouts/03_admin/03_categorization/03_categories/01_add_category/page-definition.json
@@ -4,7 +4,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.EditCategoryFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.EditCategoryJSPSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
