+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layouts/03_admin/03_categorization/page-definition.json
@@ -5,7 +5,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ViewCategorizationSectionFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.ViewCategorizationJSPSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
