+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layouts/03_admin/05_picklists/page-definition.json
@@ -4,7 +4,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.PicklistBuilderFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.PicklistBuilderComponentSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
