+++ modules/apps/site-initializer/site-initializer-cms/src/main/resources/site-initializer/layouts/05_new_space/page-definition.json
@@ -5,7 +5,7 @@
 			{
 				"definition": {
 					"fragment": {
-						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.NewSpaceFragmentRenderer"
+						"key": "com.liferay.site.cms.site.initializer.internal.fragment.renderer.NewSpaceComponentSectionFragmentRenderer"
 					},
 					"fragmentConfig": {
 					},
