+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/EditCategoryJSPSectionFragmentRenderer.java
@@ -6,11 +6,10 @@
 
 package com.liferay.site.cms.site.initializer.internal.fragment.renderer;
 
-import com.liferay.asset.kernel.service.AssetVocabularyLocalService;
 import com.liferay.fragment.renderer.FragmentRenderer;
 import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.util.Portal;
-import com.liferay.site.cms.site.initializer.internal.display.context.ViewCategoriesDisplayContext;
+import com.liferay.site.cms.site.initializer.internal.display.context.EditCategoryDisplayContext;
 
 import jakarta.servlet.http.HttpServletRequest;
 
@@ -21,30 +20,21 @@ import org.osgi.service.component.annotations.Reference;
  * @author Cheryl Tang
  */
 @Component(service = FragmentRenderer.class)
-public class ViewCategoriesFragmentRenderer
-	extends BaseJSPSectionFragmentRenderer<ViewCategoriesDisplayContext> {
+public class EditCategoryJSPSectionFragmentRenderer
+	extends BaseJSPSectionFragmentRenderer<EditCategoryDisplayContext> {
 
 	@Override
 	public String getLabelKey() {
-		return "categories";
+		return "edit-category";
 	}
 
 	@Override
-	protected ViewCategoriesDisplayContext getDisplayContext(
+	protected EditCategoryDisplayContext getDisplayContext(
 		HttpServletRequest httpServletRequest) {
 
-		return new ViewCategoriesDisplayContext(
-			_assetVocabularyLocalService, httpServletRequest,
-			_layoutLocalService, language, _portal);
+		return new EditCategoryDisplayContext(
+			httpServletRequest, _layoutLocalService, language, _portal);
 	}
-
-	@Override
-	protected String getJSPPath() {
-		return "/view_categories.jsp";
-	}
-
-	@Reference
-	private AssetVocabularyLocalService _assetVocabularyLocalService;
 
 	@Reference
 	private LayoutLocalService _layoutLocalService;
