+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/NewSpaceComponentSectionFragmentRenderer.java
@@ -22,7 +22,7 @@ import org.osgi.service.component.annotations.Component;
  * @author Roberto Díaz
  */
 @Component(service = FragmentRenderer.class)
-public class NewSpaceFragmentRenderer
+public class NewSpaceComponentSectionFragmentRenderer
 	extends BaseComponentSectionFragmentRenderer {
 
 	@Override
