+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/PicklistBuilderComponentSectionFragmentRenderer.java
@@ -22,7 +22,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Verónica Gonzaléz
  */
 @Component(service = FragmentRenderer.class)
-public class PicklistBuilderFragmentRenderer
+public class PicklistBuilderComponentSectionFragmentRenderer
 	extends BaseComponentSectionFragmentRenderer {
 
 	@Override
