+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/SpaceListComponentSectionFragmentRenderer.java
@@ -26,7 +26,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Georgel Pop
  */
 @Component(service = FragmentRenderer.class)
-public class SpaceListFragmentRenderer
+public class SpaceListComponentSectionFragmentRenderer
 	extends BaseComponentSectionFragmentRenderer {
 
 	@Override
