+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/ViewAllJSPSectionFragmentRenderer.java
@@ -21,7 +21,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Jürgen Kappler
  */
 @Component(service = FragmentRenderer.class)
-public class ViewAllSectionFragmentRenderer
+public class ViewAllJSPSectionFragmentRenderer
 	extends BaseJSPSectionFragmentRenderer<ViewAllSectionDisplayContext> {
 
 	@Override
