+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/ViewCategoriesJSPSectionFragmentRenderer.java
@@ -6,10 +6,11 @@
 
 package com.liferay.site.cms.site.initializer.internal.fragment.renderer;
 
+import com.liferay.asset.kernel.service.AssetVocabularyLocalService;
 import com.liferay.fragment.renderer.FragmentRenderer;
 import com.liferay.portal.kernel.service.LayoutLocalService;
 import com.liferay.portal.kernel.util.Portal;
-import com.liferay.site.cms.site.initializer.internal.display.context.EditCategoryDisplayContext;
+import com.liferay.site.cms.site.initializer.internal.display.context.ViewCategoriesDisplayContext;
 
 import jakarta.servlet.http.HttpServletRequest;
 
@@ -20,21 +21,30 @@ import org.osgi.service.component.annotations.Reference;
  * @author Cheryl Tang
  */
 @Component(service = FragmentRenderer.class)
-public class EditCategoryFragmentRenderer
-	extends BaseJSPSectionFragmentRenderer<EditCategoryDisplayContext> {
+public class ViewCategoriesJSPSectionFragmentRenderer
+	extends BaseJSPSectionFragmentRenderer<ViewCategoriesDisplayContext> {
 
 	@Override
 	public String getLabelKey() {
-		return "edit-category";
+		return "categories";
 	}
 
 	@Override
-	protected EditCategoryDisplayContext getDisplayContext(
+	protected ViewCategoriesDisplayContext getDisplayContext(
 		HttpServletRequest httpServletRequest) {
 
-		return new EditCategoryDisplayContext(
-			httpServletRequest, _layoutLocalService, language, _portal);
+		return new ViewCategoriesDisplayContext(
+			_assetVocabularyLocalService, httpServletRequest,
+			_layoutLocalService, language, _portal);
 	}
+
+	@Override
+	protected String getJSPPath() {
+		return "/view_categories.jsp";
+	}
+
+	@Reference
+	private AssetVocabularyLocalService _assetVocabularyLocalService;
 
 	@Reference
 	private LayoutLocalService _layoutLocalService;
