+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/ViewCategorizationJSPSectionFragmentRenderer.java
@@ -27,7 +27,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Sam Ziemer
  */
 @Component(service = FragmentRenderer.class)
-public class ViewCategorizationSectionFragmentRenderer
+public class ViewCategorizationJSPSectionFragmentRenderer
 	extends BaseSectionFragmentRenderer {
 
 	@Override
