+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/ViewContentsJSPSectionFragmentRenderer.java
@@ -21,7 +21,7 @@ import org.osgi.service.component.annotations.Reference;
  * @author Sam Ziemer
  */
 @Component(service = FragmentRenderer.class)
-public class ViewContentsSectionFragmentRenderer
+public class ViewContentsJSPSectionFragmentRenderer
 	extends BaseJSPSectionFragmentRenderer<ViewContentsSectionDisplayContext> {
 
 	@Override
