+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/fragment/renderer/ViewFilesJSPSectionFragmentRenderer.java
@@ -9,10 +9,8 @@ import com.liferay.depot.service.DepotEntryLocalService;
 import com.liferay.fragment.renderer.FragmentRenderer;
 import com.liferay.object.service.ObjectDefinitionService;
 import com.liferay.object.service.ObjectDefinitionSettingLocalService;
-import com.liferay.object.service.ObjectEntryFolderLocalService;
-import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.util.Portal;
-import com.liferay.site.cms.site.initializer.internal.display.context.ViewFolderDisplayContext;
+import com.liferay.site.cms.site.initializer.internal.display.context.ViewFilesSectionDisplayContext;
 
 import jakarta.servlet.http.HttpServletRequest;
 
@@ -20,11 +18,11 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Marco Galluzzi
+ * @author Sam Ziemer
  */
 @Component(service = FragmentRenderer.class)
-public class ViewFolderFragmentRenderer
-	extends BaseJSPSectionFragmentRenderer<ViewFolderDisplayContext> {
+public class ViewFilesJSPSectionFragmentRenderer
+	extends BaseJSPSectionFragmentRenderer<ViewFilesSectionDisplayContext> {
 
 	@Override
 	public String getCollectionKey() {
@@ -32,26 +30,27 @@ public class ViewFolderFragmentRenderer
 	}
 
 	@Override
-	protected ViewFolderDisplayContext getDisplayContext(
-		HttpServletRequest httpServletRequest) {
-
-		return new ViewFolderDisplayContext(
-			_depotEntryLocalService, _groupLocalService, httpServletRequest,
-			language, _objectDefinitionService,
-			_objectDefinitionSettingLocalService,
-			_objectEntryFolderLocalService, _portal);
+	public String getLabelKey() {
+		return "files";
 	}
 
 	@Override
-	protected String getLabelKey() {
-		return "view-folder";
+	protected ViewFilesSectionDisplayContext getDisplayContext(
+		HttpServletRequest httpServletRequest) {
+
+		return new ViewFilesSectionDisplayContext(
+			_depotEntryLocalService, groupLocalService, httpServletRequest,
+			language, _objectDefinitionService,
+			_objectDefinitionSettingLocalService, _portal);
+	}
+
+	@Override
+	protected String getJSPPath() {
+		return "/view_files_section.jsp";
 	}
 
 	@Reference
 	private DepotEntryLocalService _depotEntryLocalService;
-
-	@Reference
-	private GroupLocalService _groupLocalService;
 
 	@Reference
 	private ObjectDefinitionService _objectDefinitionService;
@@ -59,9 +58,6 @@ public class ViewFolderFragmentRenderer
 	@Reference
 	private ObjectDefinitionSettingLocalService
 		_objectDefinitionSettingLocalService;
-
-	@Reference
-	private ObjectEntryFolderLocalService _objectEntryFolderLocalService;
 
 	@Reference
 	private Portal _portal;
